Add a compile flag for no string support
Instead of having #ifdefs with __HOSTBOOT_MODULE in common code, this
flag is added to declare a policy based way of excluding code from
Hostboot that we don't support. Since Bootloader and Hostboot don't
support std::string they both declare this flag instead of bloating
common code with both __HOSTBOOT_MODULE and BOOTLOADER #ifdefs.

This is a standalone commit created to go in before policy based changes
in fapi2 so that CI can pass on that commit.

Change-Id: Iff58cf50451a7a7b6807b7b3459e1b4a1264ee66
RTC:248555
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/92142
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>