SPI Device Driver: Compile FAPI2 into Bootloader
Makes the necessary changes to compile FAPI2 into the bootloader so that
the SPI hardware procedures are able to be executed in the bootloader.
This will allow the SPI Driver code to remain commonized between
hostboot, bootloader, and SBE.

Additionally, adds policy based compile flags to common code so that
hostboot and bootloader specific preprocessor directives don't pollute the code.

Change-Id: Iaf49b120fa63ac2c8447c83047f79515c674983a
RTC:248555
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87725
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>