allow "hb workon" for non-pool machines
This allows developers to edit files locally instead of on a pool
machine. Of course developers will still be dependent on the pool
machines to build and test their code, but at least they will be able
to utilize tools like the commit hooks on their local machine.

Note that this requires NO_BUILD_TOOLS to be set to any value except
null in the customrc file.

Change-Id: Ifb699ada4b9b36bd587d995ab3b2228b5ef2f511
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/92380
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>