Do not truncate error log at runtime when using firmware_request
There is an implicit truncation in the path where we use
firmware_request to send an error log down to the FSP.  This
results in us sending a log full of zeros down instead of a
complete log because we didn't properly handle the flatten
error.  We were missing the argument to truncate, which is
required for both paths we use to send error logs.

Change-Id: Ie1a82841b33ee3d28f65d99f5707e249f2ca5a19
CQ: SW486880
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/93053
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>