Explorer update hash fix
Need to adjust what image data to perform the SHA512 hash against. Also
verify this calculated hash value matches the value included in the
image fingerprint when building the PNOR header for this image.  Need to
also fix the byte order of scom register data when grabbing the current
Explorer hash.

From vendor:
At offset 0, is the Image fingerprint which is hash function
calculated from Byte 640(dec) till the end of the FW image.

RTC:244420
Change-Id: Iae083ee7118b69794f41d54a6f52acbb42cd913b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/93698
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>