Leave more content around during PNOR build for other uses
In order to create separate lids for the content and the signature
we need to leave a few more artifacts of our PNOR build process
laying around to consume later on.  This change will cause the
signature data to be part of the saved output along with the
staged content.

Change-Id: If359a48cf05a26cf613332673d4c454f75e42e1c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/93880
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>