Add New HOMER Map/Unmap Interfaces
This commit adds a new ScopedHomerMapper class that maps HOMER physical
addresses to virtual space. The mapping occurs in the map() class
member. The HOMER virtual space is automatically deallocated on PHYP
systems when the class goes out of scope (via a class destructor). On
OPAL systems deallocation does not occur, since there is no limitation as to
the number of simultaneously mapped HOMERs.

Change-Id: I6e367d9ef7b2e9a9f05981c82d5238da9fc140fe
RTC: 251393
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94095
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christopher J Cain <cjcain@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>