FAPI2: Add wrapper boilerplate code for easy wrapper writing
Most procedure wrappers do basically the same: Loop over targets of
some kind and invoke a procedure with that target as parameter. That
should not take 200 lines of code every time :)

Replace this by a set of wrapper boilerplate includes that takes care
of everything a simple wrapper does and is customizable enough for
most practical applications:
  * Procedure name (duh!)
  * Target type (both ecmd and FAPI target)
  * Procedure invocation (if the HWP takes more than just a target)
  * Additional command line arguments (to customize the procedure)

As a proof of concept, convert most of the perv wrappers to this
style.

Change-Id: I22733406f1f3c9143bc6c8e9f09f871176690060
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94007
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt K Light <mklight@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94048
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>