Reserve correct range for Hostboot runtime data
All of the Hostboot reserved memory must be included inside the
initial "hb-rsv-mem" section.  Our move of the Hostboot HRMOR
from 128MB (0x8000000) to 4GB-256MB (0xF0000000) exposed a bug
in how we calculated the area.

Note that this current change does not correctly reserve the
data that lives below the HRMOR, which is required for MPIPL
to work.  However, the plan is to move this data into the upper
256MB range before that matters.

Change-Id: I3eadd945cf1a0ec09743d064f4a6384088eee225
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94330
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>