Add error handling for PLDM Request to MCTP Outbound Q transaction
After a PLDM request has been fully created we send it to the MCTP
Resource Provider's outbound message queue. A few things could go
wrong here. It is possible that an error could occur in the MCTP
core logic, which packetizes and writes the message onto the LPC
bus, and while we could just wait for the PLDM req to timeout, it
would be better to know about the failure right away. Also the
msg_sendrecv could fail, in which case we want to capture the return
code from that system call.

Change-Id: I222c329dfefd0a1c2222ef37850a4c5e339de8ab
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94332
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>