Invoke slave core workaround for istep 16.1/16.2 only when QME disabled
It is possible to start SIMICS with the QME model disabled. If and when that
happens, Hostboot can only pass through istep 16.1 and 16.2 successfully if
its slave core workaround is invoked.  This commit dynamically queries down into
SIMICS to determine if the QME model is disabled, and if so invokes the slave
core workaround.  Additionally, the current MSR testcase assumes SMF MSR bit
will be asserted on the thread that runs the test, however when SMF_CONFIG is 0
going into istep 16, the cores will wake up with MSR SMF bits deasserted.  This
commit fixes the testcase to consider the expected SMF_CONFIG mode when
validating the MSR, with the notable current exception that if the slave core
workaround is invoked, it assumes that SMF bit will come up as set.

Change-Id: I41f2d2d83a23909ec590dc789351bb3e785443ff
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94905
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: CALEB SCHLOSSIN <calebs@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>