Use logical I2C engine/port to compute bus speed
This commit fixes a bug in the code that computes the bus speed for a given
port, in some cases, via a non-logical engine/port combo.  This caused the same
question .. "is there a device hanging off this engine" .. to be answered
differently when the FSI path was used to reset the I2C engines, leading to
assertions.  Always checking the bus speed using the logical engine/port
solves this problem.

Change-Id: I9460ab83e3a44b7b695457678861005049350dc5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94833
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>