API for finding the FRU Record Set ID of an entity type
This commit adds a function to the PDR manager to locate the FRU
Record Set ID for a given entity type (to be used to locate the
backplane's FRU Record Set ID).

This commit also adds build flags (CONFIG_PLDM_TESTCASES and
CONFIG_MCTP_TESTCASES) that will cause libpldm and libmctp to be built
and linked, even though the IPL will not use any of their code. This
is to allow PLDM/MCTP testcases to run that don't depend on generating
MCTP traffic.

Change-Id: I2bcb543730c12f1257db3443ad20f1c48868edbe
RTC: 252496
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94974
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>