Add compile-time check to trace macros to catch using too many arguments
The trace infrastructure in Hostboot, and in particular the
TRACE::trace_adal_write_all function, which saves trace messages
created with the TRACFCOMP and TRACSCOMP macros, enforces a maximum
format argument limit (shared with fsp-trace) of TRAC_MAX_ARGS
arguments (currently 9). When more than this number of arguments are
used, the routine silently fails and the trace message is discarded.

This commit adds a compile-time check to the TRACFCOMP and TRACSCOMP
macros which will cause a compilation failure when more than
TRAC_MAX_ARGS are used. It also fixes some cases where too many
arguments were used in Hostboot code.

Because of time constraints, this commit does NOT fix the instances of
ekb code that violate the maximum argument limit. A workaround is used
for the macros FAPI_INFO, FAPI_ASSERT etc. so that they bypass the
compile-time check.

Change-Id: I96b04fb6c98ef9c3ef78b0032338a530bd53346d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/95281
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>