Print out the fields of user_input_msdg for debug
When comparing values of these fields from Microchip or from the UART,
the data is normally given in this format. Print these fields directly
in the trace for easy comparison rather than requiring parsing of the
values written out in the packed struct.

Output formats are specifically chosen to match the format used by the
UART.

Change-Id: I94fa074172648565c56eef3d0f884fa647c0fe00
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/95486
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/95548
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>