Initialize lcov coverage report with all-zero data
Currently, when we collect GCOV coverage data to build a code coverag
report, we only report trace data collected from code that actually
ran. When we see a file missing from the coverage report, this leaves
us wondering whether we did not instrument the code at all, or whethre
the code was simply not executed during our profiled run.

This commit addresses the issue by pre-running lcov using an option to
generate an all-zero coverage report, which is then augmented with the
real coverage data in a second step.

Change-Id: I582837dd22d77bc540442fbebddd4f8137903c6d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/95706
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>