Enable SBE Simics seeprom versioning
Enable the special handling to skip running the SBE seeprom updates.

For HW builds running in simics which derive their XML configs
from the MRW (or inherit global attributes) we need the ability
to flag the updateProcessorSbeSeeproms logic the ability to skip
performing the updating and writing of a new seeprom image
(to either save time during simulation or to avoid running
image customizations for a variety of reasons).

Change-Id: I471a2a8abff7e06d5b7fbb3d04211cc881fcf139
RTC: 208838
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/95748
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>