Fix code coverage relative path conversion
Code coverage requires that all file and include paths be
absolute. The script to do this had a bug that involved redundantly
checking a non-exported environment variable, so the check would
always fail and the paths would remain unconverted, causing lcov/gcov
to be unable to locate source files.

This commit fixes the problem by removing the check.

Change-Id: Ied9194325776d1b6918afbce45e680b5995e435d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96013
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>