Temporary fix for linker crashes
* Background

Hostboot contains many debug messages. The strings in these debug
messages take up considerable space in the final compiled binaries
(over 1MB in hbicore_extended alone). To reduce the size overhead of
the debug messages, a program named tracehash is used. After the
object files have been produced by the compiler, tracehash (1) removes
the trace strings and the ELF sections they are in from the object
files; (2) adds new sections to contain the "reduced format strings",
which are the format strings with all the characters except the format
specifiers removed; (3) replaces the old relocations with relocations
to these new strings; and (4) inserts hash codes for each trace
string. These hash codes are then logged to storage instead of the
full debug strings, and later the strings can be reconstructed by
referencing the hash code lookup table. The lookup table does not need
to be stored on the host, saving space.

* Problem

Hostboot recently suffered from a crash in the GNU Linker. This was
found to be related to the changes made to the object files by the
tracehash program. tracehash uses libbfd to read and modify the object
files, which is an implementation library for the GNU Compiler
Collection, and as such is poorly documented, and often changes in
backwards-incompatible ways. Along the way in Hostboot's upgrade path
through different compiler versions, new constraints, unknown to us as
yet, were placed on the structure of the object files by the GNU
Linker, and tracehash was not updated to reflect these new
constraints. This constraint violation was recently exposed by
presumably benign code changes in Hostboot. (It is also possible that
these constraints always existed, but that we only recently exposed
them for whatever reason.) So the linker now crashes when fed object
files produced by tracehash.

* Solution

Rather than discover what the problems with tracehash are (and
possibly have to repeat the process when we upgrade compilers or come
across another bug), we decided to try to implement trace hashing in
the compiler itself using C++ language features. This way, we do not
need to modify the object files using libbfd, and we can discard the
full trace string sections using a command in the linker script, which
is safer than preforming the object file surgery ourselves. The
tracehash program has been reworked into "extracthash", which only
reads the object files to extract the full trace strings and write
them to the *.o.hash files. This should work in perpetuity since this
aspect of the ELF format is not subject to change.

This commit is an implementation of this idea.

* Problem with the solution

This solution is not a panacea, however, because implementing the
string hashing and reduction in the compiler incurs a significant
compilation performance hit. Building from scratch took about 5.5
minutes before, and now takes about 8 minutes. This is unacceptable in
the long term, but for now we need to get other commits through
CI. This commit will eventually be reverted when we have a solution
that does not compromise build time.

Change-Id: Id0e42301503c34b079343353456432ac4bc5bfdc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96238
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>