Fix line breaks and a few comments  in mctp resource provider
This commit adjusts where we have the line breaks in a few places
to get us under the 80 char limit. There are a few exceptions for
things like strings where its useful for search purposes to have
the entire string on one line. Also this commit updates 2 comments
about the KCS status polling loop. Originally we were polling once
every 100 ms and we have updated that to be 1 ms in a previous commit,
this commit updates the comments to reflect this 1 ms timing also

Change-Id: Iebe055c13da17e3547aa42a46ccfb604d0cf6562
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96300
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>