Fix ATTR_SMF_CONFIG on FSP based systems
There was an issue in istep 15.1 where FSP would go down the SMF enabled
path when it should not. This happened because, for FSP builds,
distributeSmfMemory() was a noop. To correct the issue in 15.1
distributeSmfMemory() will now call distributeSmfMem() even on FSP
builds which will cause ATTR_SMF_CONFIG to be switched off during istep
7 since ATTR_SMF_MEM_REQUESTED defaults to 0.

Change-Id: Ia13cacbc50d3c132b4dacec38e4f2084401948da
RTC:252936
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96387
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>