Fix bug in processMrw to use PIB I2C connection only
The processMrw.pl script is responsible for setting the EEPROM VPD info
for the DIMMs. There was an issue where it would just take the first
I2C connection it found going out to the DIMMs and grab the I2C info to
access the EEPROM off that connection. By chance that worked in the
past, but with Denali, it has a CFAM and a PIB connnection. This commit
adds support to handle that properly.

Change-Id: I606aeb06b232e058d74547ac06500eec63487a3b
RTC:252936
CMVC-Prereq:1106146
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96572
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>