Fix MRW HUID for PMIC, OCMB, MEMPORT
Currently the HUIDs for PMIC, OCMB, and MEMPORT are incorrect.
The setHuid function bases the HUIDs on the order the targets
are parsed, but sometimes the targets are parsed in the wrong order.
We pass in a new parameter to just base the HUID on the position
of the target.

Change-Id: If9f33cb806f98d3d44d8fd152a3fec0ebbd60af4
CQ:SW491339
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96887
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>