Only check partial-good rules for children of PROC targets
This commit causes the presence detection code to avoid referring to
partial-good rules for targets that are not children of PROC
targets. This makes sense because partial-good data resides on MVPD
and is only relevant to Processor targets and their children.

This fixes an issue that the FSP team was having, which was that we
did not have PG rules defined for FSP targets and logged errors.

Change-Id: Id486fa0fc3aae3bd91f57b4b94dcfb5adca49037
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/97182
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>