Add support for handling PLDM requests from the BMC during HBRT
This commit adds suppor to rt_fwnotify so HBRT will react when the
hypervisor sends a MCTP_AVAILABLE message. Upon taking the
MCTP_AVAILABLE message, HBRT will check it has any PLDM requests to
handle and if it does it will handle it. If it does not then HBRT
will ask the hypervisor for MCTP packets until we get a PLDM request.
In the event that the hypervisor tells us there is MCTP_AVAILABLE and
we find no packets ready, hbrt the MCTP_AVAILABLE message will result
in a no-op for HBRT.

Change-Id: Ic90393bf02977c2286565600305a9bbafb9bf1fb
RTC: 249470
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/97644
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>