Remove workarounds for PDR exchange, fix PLDM entity ID assignment
This commit enables the entire BMC PDR exchange flow.

It also fixes a problem in the PLDM Entity ID assignment for
Targets. The problem was that the code did not allow the BMC to report
a FRU Record Set ID count of zero for any entity type, which is
incorrect, because we do not instantiate any TPM targets yet.

Change-Id: I534689690b8e6c8e9dcaacca4479b0ba65d78fb2
RTC: 252081
RTC: 249981
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/97936
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>