ScopedHomerMapper: FSP Fix
ScopedHomerMapper had a bug where it always relied on
get_pm_complex_addresses runtime interface when doing the map
of HOMER/OCC Common addresses in runtime. While that is acceptable on eBMC
systems, on FSP systems the addresses are provided to us in the function
call to load and start OCC, and ScopedHomerMapper mustn't rely on that
interface being available. This commit introduces FSP-specific logic to
handle this scenario by reading the physical addresses of HOMER/OCC
common from the appropriate attributes in FSP case.

Change-Id: Ibc09fd55819ab71a7d350c51b653f7d3bfce6b38
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98044
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>