Only check MC_SYNC_MODE for SCOM multicast workaround on Nimbus
This commit fixes a problem with the SCOM multicast workaround, which
is that it was checking the MC_SYNC_MODE on all models rather than
just Nimbus for which the attribute was intended. This was causing
issues in the multicast SCOM request to collect a bit from
MC_CHIPLET_UCS_FIR from each MC and put it in GLOBAL_UCS_FIR when a
unit checkstop occurs. Rather than collect the bits as intended, the
SCOM read was returning 0.

Change-Id: Ie20148d1be285d664d11c2ea081b55d80d85b9bd
CQ: SW492080
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98192
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>