vbu/contained: Randomize and set thread seeds for exercisers
Exercisers use the SIER register(s) to seed the current testcase.
Implement support for specifying a "master" seed to generate random
values for each core's threads via the C++11 Mersenne Twister engine.

Default randomization occurs as follows:

	1. ATTR_RUNN_MASTER_SEED sets the starting seed of the RNG
	2. For each core, generate 4 new random values from the RNG, one
	   per thread (this advances the internal RNG state)

Overrides are available:

	1. Set the per-core ATTR_RUNN_THREAD_SEEDS attribute manually in
	   an attribute seedfile (qualify with `CONST` to override) to
	   bypass the RNG completely
	2. Set ATTR_RUNN_MASTER_SEED and set ATTR_RUNN_CORE_SEED_SELECT
	   to use the RNG starting point of a specific core for all
	   cores (this effectively means: given a master-seed, duplicate
	   a single core's thread seeds across all others using the RNG.
	   This mode is intended for repro-cases).

This new randomization is qualified with a compile time define,
`P10_CONTAINED_ENABLE_SEEDING`, to allow platforms without C++11
`random` support to disable this feature.

Also fix ringspinner to use the Spy's actual latchname when constructing
the inversion mask. Previously, the latchname was inferred from the
Spyname which is not always correct. As it turns out, the SIER Spynames
are completely different from their latchnames and thus exposed this bug.

Change-Id: I3ee8478f993936f2dac4da862102194b6faf3418
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98221
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98286
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>