Update hb prime to automatically build hostboot patches for FSP
Updates the hb prime script to automatically build the hostboot
patches that belong in FSP's /nfs/test/pnor/ directory. Additionally,
prime will copy the patches to $HB_NFS_DIR/test/pnor if defined or to the
$SANDBOXBASE/nfs/test/pnor/ directory it makes. This will allow
startup.simics to kick off an autoboot and automatically patch the FSP
for the user in an upcoming commit.

Change-Id: I3cfed995547b72ba6670579b5693065c23f65454
RTC:246815
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/98942
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>