Adding attributes to disable L2 & L3 hashing
Several exerciser count on cache hashing to be diabled for rolling
of the L2 & L3 caches. These dials are in the L2 & L3 scoms, so setting
them post IPL via scom would be reset when the caches go to sleep. The
solution was to add an attribute to disable hashing that could be set
by exercisers.

Change-Id: Iac318e042ca5c14df44d7fb4b0b71a731ba539c2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99147
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99273
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>