Use old version of SPI code
There is an incompatibility between both the older
SPI code with the new newer SPI model, and the
newer SPI code with the older SPI model.  This
change brings in a separate copy of the HWP with
the current (soon to be old) content.  Simics will
provide a variable (spi_master_version) that tells
us which version of the model is in use.  We will
choose the appropriate version of the HWP based
on this variable (check is via MAGIC instruction).

Change-Id: I5f20e5eea81f68a5c0e2284665f2bd8440e295dc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99258
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>