Ensure at least 8 MB of cache space is setup during kernel init
During the page manager init we will dcbz cache lines to prepare
the cache for hostboot to run. Previously we were always using the
size of avail cache we were give by the SBE. In an MPIPL however,
the SBE will return 0 for this size. In this case we want to revert
to a valid default.

Change-Id: I6a1c8aa2fd8d2f8ccdd068bf495a808614d638a9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99424
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>